requirements.txt
Django
django-filter
djangorestframework
joblib
Markdown
numpy
pandas
requests
scikit-learn
